game_bot/requirements.txt
python-telegram-bot>=20.0
grpcio>=1.32.0
protobuf>=3.13.0